refactor(src): Impose module structure (matching CLI functions)
