+++ src/inference/inference.py
@@ -1,7 +1,7 @@
 import torch
 
 from src.dataclass import Context
-from src.utils import encode, decode
+from src.utils.utils import encode, decode
 
 
 def complete(ctx: Context, model: torch.nn.Module, prompt: str, temperature: float, generated_tokens: int) -> str:
